Working on displaying data, fetching data for multiple apps at once
In frontend's App.js, continued working on components that display the API's data. In steamApps.js, added option for getting data for multiple apps at once (Steam API doesn't appear to support that directly, so instead multiple requests for individual apps are made and the response data for each is added to an array).