Fetching app news, more components
In the backend, added possibility to fetch latest news for specified app; also added a function for this in frontend's steamApps.js; not currently in use.

Continued refactoring App.js by creating separate file for SteamAppDetailed. Related to this, created ImageGrid.js for displaying screenshots of a given app.

In SteamAppSummaryList, added possibility to select an app to display by clicking it in the list.

Removed some console loggers.